+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 1.2.0
 
 * First release.
+* List users assigned to a given role. Refs UIROLES-31.